2021.05.19 最终版提交

+++ src/main/java/com/bookstore/haid/model/ShopCart.java
@@ -2,6 +2,8 @@ package com.bookstore.haid.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ShopCart {
     private Integer id;
@@ -9,7 +11,7 @@ public class ShopCart {
     private String userName;
     private String bookName;
     private String imgUrl;
-    private Double bookPrice;
+    private BigDecimal bookPrice;
     private Integer buyNum=1;
-    private Double subTotal;
+    private BigDecimal subTotal;
 }
